pyproject.toml
[project]
name = "vpn-link-converter"
version = "0.1.0"
description = "Telegram bot and web service for converting VPN links to sing-box configuration"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "jinja2>=3.1.2",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.24.0",
    "pyyaml>=6.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.scripts]
vpn-convert = "cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

